feat(cell::PredCell): Axiomatise PredCell from RustHornBelt development
