+++ creusot-contracts/src/lib.rs
@@ -395,6 +395,7 @@ pub mod logic;
 pub mod model;
 pub mod pcell;
 pub mod peano;
+pub mod predcell;
 pub mod ptr_own;
 pub mod resolve;
 pub mod resource;
