+++ tests/should_fail/diagnostics/view_unimplemented.stderr
@@ -41,7 +41,7 @@ error[E0277]: Cannot take the model of `S`
              creusot_contracts::Ghost<T>
              creusot_contracts::Snapshot<T>
              creusot_contracts::Subset<T>
-           and 49 others
+           and 50 others
 
 error[E0277]: Cannot take the model of `S`
   --> view_unimplemented.rs:14:25
@@ -59,7 +59,7 @@ error[E0277]: Cannot take the model of `S`
              creusot_contracts::Ghost<T>
              creusot_contracts::Snapshot<T>
              creusot_contracts::Subset<T>
-           and 49 others
+           and 50 others
 
 error: aborting due to 3 previous errors
