+++ tests/should_succeed/cell/01_basic.coma
@@ -1,10 +1,10 @@
-module M_01_basic__adds_two [#"01_basic.rs" 33 0 33 34]
-  let%span s01_basic = "01_basic.rs" 37 11 37 17
-  let%span s01_basic'0 = "01_basic.rs" 38 18 38 19
-  let%span s01_basic'1 = "01_basic.rs" 40 14 40 15
-  let%span s01_basic'2 = "01_basic.rs" 32 11 32 43
-  let%span s01_basic'3 = "01_basic.rs" 20 4 20 29
-  let%span s01_basic'4 = "01_basic.rs" 26 4 26 25
+module M_01_basic__adds_two [#"01_basic.rs" 6 0 6 34]
+  let%span s01_basic = "01_basic.rs" 10 11 10 17
+  let%span s01_basic'0 = "01_basic.rs" 11 18 11 19
+  let%span s01_basic'1 = "01_basic.rs" 13 14 13 15
+  let%span s01_basic'2 = "01_basic.rs" 5 11 5 37
+  let%span spredcell = "../../../creusot-contracts/src/predcell.rs" 65 4 65 29
+  let%span spredcell'0 = "../../../creusot-contracts/src/predcell.rs" 34 4 34 25
   let%span smodel = "../../../creusot-contracts/src/model.rs" 44 8 44 22
   let%span smapping = "../../../creusot-contracts/src/logic/mapping.rs" 67 4 67 12
   
@@ -17,22 +17,22 @@ module M_01_basic__adds_two [#"01_basic.rs" 33 0 33 34]
   
   predicate index_logic (self: Map.map UInt32.t bool) (a: UInt32.t) = [%#smapping] Map.get self a
   
-  function view [#"01_basic.rs" 13 4 13 33] (self: t_PredCell) : Map.map UInt32.t bool
+  function view (self: t_PredCell) : Map.map UInt32.t bool
   
   function view'0 (self: t_PredCell) : Map.map UInt32.t bool = [%#smodel] view self
   
   let rec get (self: t_PredCell) (return' (x: UInt32.t)) = any
-    [ return''0 (result: UInt32.t) -> {[%#s01_basic'3] index_logic (view'0 self) result} (! return' {result}) ]
+    [ return''0 (result: UInt32.t) -> {[%#spredcell] index_logic (view'0 self) result} (! return' {result}) ]
   
   let rec set (self: t_PredCell) (v: UInt32.t) (return' (x: ())) =
-    {[@expl:set requires] [%#s01_basic'4] index_logic (view'0 self) v}
+    {[@expl:set requires] [%#spredcell'0] index_logic (view'0 self) v}
     any [ return''0 (result: ()) -> (! return' {result}) ]
   
   meta "compute_max_steps" 1000000
   
   meta "select_lsinst" "all"
   
-  let rec adds_two [#"01_basic.rs" 33 0 33 34] (c: t_PredCell) (return' (x: ())) =
+  let rec adds_two [#"01_basic.rs" 6 0 6 34] (c: t_PredCell) (return' (x: ())) =
     {[@expl:adds_two requires] [%#s01_basic'2] view'0 c = fun (z: UInt32.t) -> Int.mod (UInt32.t'int z) 2 = 0}
     (! bb0
     [ bb0 = s0 [ s0 = get {c'0} (fun (_ret: UInt32.t) -> [ &v <- _ret ] s1) | s1 = bb1 ]
