+++ tests/should_succeed/cell/02_fib.rs
@@ -1,33 +1,6 @@
 extern crate creusot_contracts;
 
-use creusot_contracts::{logic::Mapping, *};
-
-#[trusted]
-pub struct PredCell<T>(std::cell::Cell<T>);
-
-impl<T> View for PredCell<T> {
-    type ViewTy = Mapping<T, bool>;
-
-    #[logic]
-    #[trusted]
-    fn view(self) -> Self::ViewTy {
-        dead
-    }
-}
-
-impl<T: Copy> PredCell<T> {
-    #[trusted]
-    #[ensures(self@[result])]
-    pub fn get(&self) -> T {
-        self.0.get()
-    }
-
-    #[trusted]
-    #[requires(self@[v])]
-    pub fn set(&self, v: T) {
-        self.0.set(v)
-    }
-}
+use creusot_contracts::{predcell::PredCell, *};
 
 #[logic]
 #[open]
